feat: move attributes and __init__ method
