+++ src/fault_injector/fault_types/platform_blocked_fault.py
@@ -2,11 +2,6 @@ from src.fault_injector.fault_configurations.platform_blocked_fault_configuratio
     PlatformBlockedFaultConfiguration,
 )
 from src.fault_injector.fault_types.fault import Fault
-from src.interlocking_component.route_controller import IInterlockingDisruptor
-from src.logger.logger import Logger
-from src.wrapper.simulation_object_updating_component import (
-    SimulationObjectUpdatingComponent,
-)
 from src.wrapper.simulation_objects import Platform
 
 
@@ -15,23 +10,6 @@ class PlatformBlockedFault(Fault):
 
     configuration: PlatformBlockedFaultConfiguration
     platform: Platform
-    wrapper: SimulationObjectUpdatingComponent
-    interlocking: IInterlockingDisruptor
-
-    # pylint: disable=duplicate-code
-    # Otherwise another inheritance layer would be needed.
-    def __init__(
-        self,
-        configuration,
-        logger: Logger,
-        wrapper: SimulationObjectUpdatingComponent,
-        interlocking: IInterlockingDisruptor,
-    ):
-        super().__init__(configuration, logger)
-        self.wrapper = wrapper
-        self.interlocking = interlocking
-
-    # pylint: enable=duplicate-code
 
     def inject_fault(self, tick: int):
         """inject PlatformBlockedFault into the given component
