+++ src/fault_injector/fault_types/train_prio_fault.py
@@ -2,11 +2,6 @@ from src.fault_injector.fault_configurations.train_prio_fault_configuration impo
     TrainPrioFaultConfiguration,
 )
 from src.fault_injector.fault_types.fault import Fault
-from src.interlocking_component.route_controller import IInterlockingDisruptor
-from src.logger.logger import Logger
-from src.wrapper.simulation_object_updating_component import (
-    SimulationObjectUpdatingComponent,
-)
 from src.wrapper.simulation_objects import Train
 
 
@@ -16,21 +11,6 @@ class TrainPrioFault(Fault):
     configuration: TrainPrioFaultConfiguration
     old_prio: int
     train: Train
-    wrapper: SimulationObjectUpdatingComponent
-    interlocking: IInterlockingDisruptor
-
-    # pylint: disable=duplicate-code
-    # Otherwise another inheritance layer would be needed. This will be refactored in the future
-    def __init__(
-        self,
-        configuration,
-        logger: Logger,
-        wrapper: SimulationObjectUpdatingComponent,
-        interlocking: IInterlockingDisruptor,
-    ):
-        super().__init__(configuration, logger)
-        self.wrapper = wrapper
-        self.interlocking = interlocking
 
     def inject_fault(self, tick: int):
         """inject TrainPrioFault into the given component
@@ -43,7 +23,6 @@ class TrainPrioFault(Fault):
             for train in self.wrapper.trains
             if train.identifier == self.configuration.affected_element_id
         ][0]
-        # pylint: enable=duplicate-code
         self.old_prio = self.train.train_type.priority
         self.train.train_type.priority = self.configuration.new_prio
